refactor: reuse the shared pctOf/memPct/diskPct in the alerts fold
From /be-review architecture-first-principles (C1, confirmed): alerts.ts had a
third copy of pctOf and re-derived mem%/disk% instead of the existing helpers.
Hoist pctOf/memPct/diskPct DOWN into the agent-shared `drishti-common/metrics`
(alongside averageCoreUsage — the one home the agent + its alerts fold import),
re-export them from the app-tier metrics module so history.ts / client keep
resolving one module, and have metricsFrameOf use memPct/diskPct. No tier
re-derives the share-of-total formula now.
